리팩토링: AITConvertCore에서 WebGL 빌드 실행부와 에러 카탈로그 분리
AITConvertCore.cs(1,235행)를 #36의 AITBuildCancellation과 동일한
파사드 위임 패턴으로 분리한다. 공개 API·에러 메시지 문자열·빌드
동작은 변경 없음.

- Editor/AITWebGLBuilder.cs 신설: ShouldForceCleanBuild, BuildWebGL,
  버전 정보 JSON 기록/정리, git 커밋 해시 조회를 이동 (internal static)
- Editor/AITExportErrorCatalog.cs 신설: GetErrorMessage /
  GetErrorShortReason / GetErrorCause 스위치 본문을 이동
  (Sentry 분류기·테스트가 참조하는 에러 문자열은 그대로 유지)
- AITConvertCore는 687행 파사드로 축소, 기존 호출부는 위임으로 유지
- BuildMarkerTests: private 리플렉션 호출을 InternalsVisibleTo 기반
  직접 호출로 전환 (AppsInTossEditModeTests)
- TODO.md: P2 잔여 2건 정리 — 본 분리 작업 완료 + Thread.Sleep
  전환 건은 #860에서 이미 완료된 항목의 잔재 제거

검증: ./run-local-tests.sh --validate 4/4 통과, EditMode 배치모드
981건 중 967 통과 / 0 실패 (BuildMarkerTests 5/5 포함)
